serverA/requirements.txt
flask==2.3.2
flask-sqlalchemy==3.0.5
flask-jwt-extended==4.7.1
flask-caching==2.3.0
flask-mailman==1.1.1
flask-cors==4.0.0
celery==5.3.0
redis==4.6.0
python-dotenv
werkzeug==3.0.0
google-generativeai